Do not delete cache (/thumbs, /slides) folders, because users may have links in posts to these images

+++ tools/class-eazyest-upgrade-engine.php
@@ -13,7 +13,7 @@ if ( !defined( 'ABSPATH' ) ) exit;
  * @author Marcel Brinkkemper
  * @copyright 2012 Brimosoft
  * @since 0.1.0 (r2)
- * @version 0.1.0 (r56)
+ * @version 0.1.0 (r64)
  * @access public
  */
 class Eazyest_Upgrade_Engine {
@@ -554,6 +554,7 @@ class Eazyest_Upgrade_Engine {
 	 * 
 	 * @since 0.1.0 (r2)
 	 * @uses get_transient()
+	 * @uses aplly_filters()) for ('eazyest_gallery_delete_cache',false) because by default chache is not deleted
 	 * @uses get_post()
 	 * @uses wp_update_post()
 	 * @uses get_post_meta()
@@ -562,11 +563,19 @@ class Eazyest_Upgrade_Engine {
 	function do_upgrade_folder() {		
 		$upgrade_folders = get_transient( 'eazyest-gallery-upgrade-folders' );
 		if ( $upgrade_folders ) {
-			$this->delete_cache();
+			// apply filter to delete cache (false)
+			if ( apply_filters( 'eazyest_gallery_delete_cache', false ) )
+				$this->delete_cache();
+			
+			// get folder path to upgrade
 			$raw_path = $upgrade_folders[0];
 			$folder_title = basename( $raw_path );
+			
+			// convert dashes and hyphens to spaces
 			if ( eazyest_folderbase()->replace_dashes() )
-				$folder_title = str_replace( array( '-', '_'), ' ', $folder_title );			
+				$folder_title = str_replace( array( '-', '_'), ' ', $folder_title );		
+			
+			// insert foldr in wpdb and retrieve stored/sanitized galery_path		
 			$folder_id = eazyest_folderbase()->insert_folder( $raw_path );
 			if ( $folder_id ) {
 				$gallery_path = get_post_meta( $folder_id, '_gallery_path', true );
@@ -579,6 +588,7 @@ class Eazyest_Upgrade_Engine {
 						$upgrade_folders[$key] = $upgrade_folder;
 					}
 				}
+				// store folders array for next AJAX call
 				set_transient( 'eazyest-gallery-upgrade-folders', $upgrade_folders, 0 );
 				
 				$post_parent = 0;
